veux: Make more GKI compatible.
- recovery_as_boot -> move_recovery_to_vendor_boot.
Recovery resources are built to vendor_boot instead
of boot ramdisk. This is needed for GKI testing because
the latest GKI does not contain device-specific files and
recovery resources any more, including fstab files.

- Set BOARD_USES_GENERIC_KERNEL_IMAGE to true so that `rw`
is added to kernel command line so that e2fsck is found in
the correct place.

Bug: 172510680
Bug: 172603832
Test: boots

Change-Id: I46f19d040d02781b283054decc60938fc736bab1